Renderer: remove duplicate @conversation interceptor
The renderer previously intercepted both '/conversation' and
'@conversation' submissions in chatView.addUserMessage and routed
them through the renderer-side handleConversationCommand helper.
That helper duplicates logic already implemented in the dispatcher
(systemAgent -> getConversationCommandHandlers -> manage-conversation
client action), and it's been a recurring source of drift — most
recently this PR had to mirror the new prev/next/cycle behaviour in
both places.

Make the dispatcher the single source of truth:

- Drop conversationCommands.ts entirely.  The dispatcher's
  ConversationSwitchCommandHandler already cycles when no name is
  given, and ConversationPrev/NextCommandHandler exist; subcommand
  results are surfaced via the manage-conversation client action
  handler in main.ts.
- chatView.addUserMessage now only normalises the legacy
  '/conversation' slash alias to '@conversation' before dispatch;
  the dispatcher handles the rest.
- Move the previously local 'escapeHtml' helper to a small
  htmlUtil.ts module so main.ts can keep using it.
